Bugfix
Sets a default canvas height so that render calculations can be
simplified. This height can be overriden from a stylesheet.
Callbacks are now handled at the change event, this is so that the
slider value can be programatically changed.
Fixed a bug when calculating the maximum height. min and max where
misplaced.
Fixed a bug where the selected bar wouldn't be highlighted when step !=
1
Labels' text can now be formatted via the label.format option